Moved Magic_Methods trait to traits/magic_methods.php
To keeep better directory structure
